Renames classes and files from Linear to metric and added additional testcases.
Changes from naming Linear to naming Metric. The metric category is going to contain linear and logarithmic axes.
(A metric axis-frame is an axisframe where you have a distance measure.)
Also updated the examples and added lines-charts and area charts.

Open issue is the horizontal orientation. There is an offset in the y-axis that needs to be removed.