[ANALYZER-2548] - Sunburst slice proportions are not equivalent to the represented values' proportions Categorical charts group datums by series and then by category. All the datums that fall in the same leaf-group are represented by a single visual-element. An aggregation of the value of each of those datums is what is actually represented. There's no way to specify the type of aggregation that is performed in CCC: it's always SUM. Until now, the UseAbs option of continuous axes could, when defined on a type of Axis, could affect this sum, by taking the ABS of each value, before summing it. In practice, only continuous SizeAxis and ColorAxis have and used this option meaningfully.
The old AngleAxis (now renamed to NormalizedAxis) had the UseAbs option (now removed!), but it was used for controlling both:
i) aggregation of datums' values falling in the same scene (taking the ABS of each datum's value), and
ii) calculating the sum of scene values into total values, for percentages calculation (taking the ABS of each scene value).

The UseAbs option was removed from the NormalizedAxis cause it would always have to be true.

For cartesian axes, percentages are also used in stacked layouts (stacked bar, area), and are even shown in the tooltip.

After this change, the aggregation of datums' values falling in the same scene is always done WITHOUT ABS.
However, the result of aggregation can still yield negative values, and so, UseAbs still has a reason to exist, although its usage has been clarified.

Analyzer doesn't use more than one datum per main visual element scene, and so is unaffected by this change.

It was quite confusing in the code that some sums were aggregating datums values for a scene and others were summing scene's values.
In the future, we'd like that the type of aggregation that is performed could be configured.
To that end and clarification, new methods were added to Dimension: value and valueAbs.
These return the aggregated value and its absolute value, of all the datums on the dimension, respectively.
Whether the actual aggregation is a sum or not gets abstracted away. The code gets easier to read as well.