[BACKLOG-9117] Automatic axis label's non-overlapping angle (#244)
* Fixed the grid layout algorithm to handle orthogonal size changes of side panels
during the first layout phase.
Refactor of docking grid panel layout code, to make it easier to reason about...
Currently, the dock layout only allows axes to change the ortho length in the
Phase 1... which reveals itself as invalid layouts for certain cases
caused by the auto-angle and the changes in ortho-length it causes.

* [BACKLOG-9117] Automatic axis label's non-overlapping angle

* [BACKLOG-9117] Documentation for axis label's non-overlapping angle

* [BACKLOG-9117] Tests for axis label's non-overlapping angle

* [BACKLOG-9117] Documentation review for labels non-overlapping angle

* [BACKLOG-9117] Moved docs to DiscreteCartesianAxisTicksFacet

* [BACKLOG-9117] Fixes after code review feedback