Adding to the existing Options JsDoc documentation the core parts related with the “Visual API” - essentially, other namespaces, such `pvc.data`, `cdo`, `def`, `pvc` and `pvc.visual`. (#325)
It is by no means complete. Notably, entities such as visual roles, plots and plot panels are not documented. Text utilities, and other top-level utility functions are also not documented. To be done.

Also, fixed an issue with the XSL that generates the `pvc.options` part of the documentation, in which enum values composed of dashes were not being generated.